World Happiness Report project from Kaggle
A little context on World Happiness Report:
The World Happiness Report is a landmark survey of the state of global happiness. The reports review the state of happiness in the world today and show how the new science of happiness explains personal and national variations in happiness. I have analyzed the report on World Happiness from the year 2017. The report ranks 155 countries by their happiness levels. The predictor variables: GDP per Capita, Family, Life Expectancy, Freedom, Generosity, Trust Government, and Corruption describe the extent to which these factors contribute in evaluating the happiness in each country. The reports review the state of happiness in the world today and show how the new science of happiness explains personal and national variations in happiness.